requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
click==8.1.7
databases==0.8.0
et-xmlfile==1.1.0
fastapi==0.104.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
qrcode==7.4.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.24.0.post1
